requirements.txt
Django==4.2.21
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-nested-routers==0.94.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
celery==5.5.3
redis==6.2.0
django-celery-beat==2.8.1
django-extensions==4.1